backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.2.0
Pillow==10.0.1
python-multipart==0.0.6
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0 